requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pandas==2.2.0
openpyxl==3.1.2
groq==0.5.0
matplotlib==3.8.3
reportlab==4.1.0